+++ partiql-app/partiql-cli/src/main/kotlin/org/partiql/cli/pico/PartiQLCommand.kt
@@ -84,11 +84,9 @@ internal class PartiQLCommand(private val ion: IonSystem) : Runnable {
             false -> query
             else -> queryLines.subList(1, queryLines.size).joinToString(System.lineSeparator())
         }
-        input.use { src ->
-            output.use { out ->
-                Cli(ion, src, exec.inputFormat, out, exec.outputFormat, options.pipeline, options.environment, queryWithoutShebang, exec.wrapIon).run()
-                out.write(System.lineSeparator().toByteArray(Charsets.UTF_8))
-            }
+        output.use { out ->
+            Cli(ion, input, exec.inputFormat, out, exec.outputFormat, options.pipeline, options.environment, queryWithoutShebang, exec.wrapIon).run()
+            out.write(System.lineSeparator().toByteArray(Charsets.UTF_8))
         }
     }