+++ partiql-app/partiql-cli/src/main/kotlin/org/partiql/cli/utils/InputSource.kt
@@ -17,14 +17,26 @@ package org.partiql.cli.utils
 import java.io.File
 import java.io.InputStream
 
+/**
+ * Represents a re-openable source containing input data. By calling [stream], the [InputSource] opens a stream at the
+ * beginning of the represented data.
+ */
 internal sealed class InputSource {
 
     internal abstract fun stream(): InputStream
 
+    /**
+     * Represents a re-openable source wrapping a file's contents. By calling [stream], the [FileSource] opens a stream
+     * at the beginning of the [file]
+     */
     internal class FileSource(private val file: File) : InputSource() {
         override fun stream(): InputStream = file.inputStream()
     }
 
+    /**
+     * Represents a re-openable source wrapping a string's contents. By calling [stream], the [StringSource] opens a stream
+     * at the beginning of the [string]
+     */
     internal class StringSource(private val string: String) : InputSource() {
         override fun stream(): InputStream = string.byteInputStream(Charsets.UTF_8)
     }